refactor: 분리된 활동 로그 컴포넌트 적용
