+++ src/pages/club-detail/TabContents.tsx
@@ -1,5 +1,5 @@
 import Intro from '@/components/ClubDetail/ClubDetailTabs/Intro/Intro';
-import Log from './tab/Log';
+import Log from '@/components/ClubDetail/ClubDetailTabs/Log/Log';
 import Recruit from '@/components/ClubDetail/ClubDetailTabs/Recruit/Recruit';
 
 interface TabContentsProps {
